+++ src/util/linear_buffer.c
@@ -14,7 +14,7 @@
 
 void linear_buffer_init(
     LinearBuffer *linear_buffer,
-    uint32_t *buffer,
+    uint16_t *buffer,
     uint32_t size
 )
 {
@@ -34,7 +34,7 @@ void linear_buffer_init(
 
 uint32_t linear_buffer_read(
     LinearBuffer *linear_buffer,
-    uint32_t *data,
+    uint16_t *data,
     uint32_t size
 )
 {
